requirements.txt
google-genai
python-dotenv
sentence-transformers
numpy
scikit-learn
